Ready for review.
HttpCrawler made generic.
BeautifulSoup and Parsel crwalers inherit from this new generic.
