+++ src/crawlee/http_crawler/__init__.py
@@ -1,4 +1,13 @@
-from ._http_crawler import HttpCrawler
-from ._http_crawling_context import HttpCrawlingContext, HttpCrawlingResult
+from ._http_crawler import HttpCrawler, HttpCrawlerGeneric
+from ._http_crawling_context import HttpCrawlingContext, HttpCrawlingResult, ParsedHttpCrawlingContext
+from ._http_parser import BlockedInfo, StaticContentParser
 
-__all__ = ['HttpCrawler', 'HttpCrawlingContext', 'HttpCrawlingResult']
+__all__ = [
+    'BlockedInfo',
+    'HttpCrawler',
+    'HttpCrawlerGeneric',
+    'HttpCrawlingContext',
+    'HttpCrawlingResult',
+    'ParsedHttpCrawlingContext',
+    'StaticContentParser',
+]
